# docs/swagger.yaml
basePath: /
definitions:
  wiki_internal_dto.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
  wiki_internal_dto.ErrorResponse:
    description: Стандартный ответ при ошибке
    properties:
      error:
        example: Описание ошибки
        type: string
    type: object
  wiki_internal_dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  wiki_internal_dto.LoginResponse:
    properties:
      access_token:
        type: string
      user:
        $ref: '#/definitions/wiki_internal_dto.UserResponse'
    type: object
  wiki_internal_dto.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  wiki_internal_dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для корпоративной Wiki (Auth, Users)
  title: Corporate Wiki API
  version: "1.0"
paths:
  /api/profile:
    get:
      consumes:
      - application/json
      description: Возвращает данные текущего пользователя по JWT-токену в заголовке
        Authorization
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wiki_internal_dto.UserResponse'
        "401":
          description: Невалидный или отсутствующий токен
          schema:
            $ref: '#/definitions/wiki_internal_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/wiki_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Профиль пользователя
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Проверяет логин и пароль, возвращает JWT-токен и данные пользователя
      parameters:
      - description: Учётные данные
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/wiki_internal_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wiki_internal_dto.LoginResponse'
        "400":
          description: Некорректный запрос (невалидный JSON)
          schema:
            $ref: '#/definitions/wiki_internal_dto.ErrorResponse'
        "401":
          description: Неверный email или пароль
          schema:
            $ref: '#/definitions/wiki_internal_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/wiki_internal_dto.ErrorResponse'
      summary: Вход в систему
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя через публичный эндпоинт регистрации
      parameters:
      - description: Данные регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/wiki_internal_dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/wiki_internal_dto.UserResponse'
        "400":
          description: Некорректный запрос (невалидный JSON)
          schema:
            $ref: '#/definitions/wiki_internal_dto.ErrorResponse'
        "409":
          description: Пользователь с таким email уже существует
          schema:
            $ref: '#/definitions/wiki_internal_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/wiki_internal_dto.ErrorResponse'
      summary: Регистрация
      tags:
      - auth
  /health:
    get:
      consumes:
      - application/json
      description: Возвращает статус приложения (alive, starting и т.д.)
      produces:
      - application/json
      responses:
        "200":
          description: Статус приложения
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Проверка состояния
      tags:
      - health
  /users:
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя в системе (административный эндпоинт)
      parameters:
      - description: Данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/wiki_internal_dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/wiki_internal_dto.UserResponse'
        "400":
          description: Некорректный запрос (невалидный JSON)
          schema:
            $ref: '#/definitions/wiki_internal_dto.ErrorResponse'
        "409":
          description: Пользователь с таким email уже существует
          schema:
            $ref: '#/definitions/wiki_internal_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/wiki_internal_dto.ErrorResponse'
      summary: Создать пользователя
      tags:
      - users
  /users/:id:
    get:
      consumes:
      - application/json
      description: Возвращает данные пользователя по его числовому идентификатору
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wiki_internal_dto.UserResponse'
        "400":
          description: Некорректный ID пользователя
          schema:
            $ref: '#/definitions/wiki_internal_dto.ErrorResponse'
        "404":
          description: Пользователь с таким ID не найден
          schema:
            $ref: '#/definitions/wiki_internal_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/wiki_internal_dto.ErrorResponse'
      summary: Получить пользователя по ID
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"